Homework: Restriction access for delete article

+++ app/controllers/articles_controller.rb
@@ -46,8 +46,13 @@ class ArticlesController < ApplicationController
   end
 
   def destroy
-    if @article.destroy
-      redirect_to articles_path, notice: 'Article has been deleted'
+    unless @article.user == current_user
+      flash[:alert] = "You can only delete your own article."
+      redirect_to root_path
+    else
+      if @article.destroy
+        redirect_to articles_path, notice: 'Article has been deleted'
+      end
     end
   end
